requirements.txt
holidays
requests
flask
kafka-python
